Transfer and separate the admin and customer APIs to the dashboard app.
