+++ app-backend/accounts/api/V1/urls.py
@@ -5,11 +5,7 @@ from .views import (
     ResendEmailView,
     ChangePasswordView,
     PasswordResetRequestView,
-    PasswordResetConfirmView,
-    GetProfileCustomerAPIView,
-    AllProfileCustomerAPIView,
-    UpdateProfileCustomerAPIView
-
+    PasswordResetConfirmView
 )
 from rest_framework_simplejwt.views import (
     TokenObtainPairView,
@@ -27,10 +23,6 @@ urlpatterns = [
     path("change-password/", ChangePasswordView.as_view(), name="change-password"),
     path("reset-password/", PasswordResetRequestView.as_view(), name="reset-password"),
     path("reset-password-confirm/<str:token>/", PasswordResetConfirmView.as_view(), name="reset-password-confirm"),
-    path("all_profile/", AllProfileCustomerAPIView.as_view(), name="all_profile"),
-    path("profile_customer/", GetProfileCustomerAPIView.as_view(), name="get_profile"),
-    path("profile_customer/<int:pk>/update/", UpdateProfileCustomerAPIView.as_view(), name="update_profile"),
-
     #simplejwt
     path('token/login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
     path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
