+++ app-backend/dashboard/customer/api/V1/urls.py
@@ -1,0 +1,11 @@
+from django.urls import path
+from dashboard.customer.api.V1.views import GetProfileCustomerAPIView, UpdateProfileCustomerAPIView, \
+                                            GetReservationCustomerListAPIView
+
+
+
+urlpatterns = [
+    path("profile_customer/", GetProfileCustomerAPIView.as_view(), name="get_profile"),
+    path("profile_customer/<int:pk>/update/", UpdateProfileCustomerAPIView.as_view(), name="update_profile"),
+    path('reserve-list_customer/', GetReservationCustomerListAPIView.as_view(), name='reserve-list-customer'),
+]
