+++ app-backend/reservations/api/V1/views.py
@@ -1,7 +1,6 @@
 from rest_framework import generics
 from reservations.models import Reservation  # دقت کنید که مسیر صحیح مدل رزرو را وارد کرده‌اید
-from .serializers import ReservationSerializer, CustomerReservationSerializer
-from rest_framework.permissions import IsAuthenticated, IsAdminUser
+from .serializers import ReservationSerializer
 from reservations.api.V1.tasks import send_reservation_email  # وارد کردن تسک به مسیر صحیح
 
 
@@ -23,24 +22,3 @@ class ReservationCreateView(generics.CreateAPIView):
         except Exception as e:
             # در صورت بروز مشکل می‌تونی پیام خطا رو چاپ کنی یا ذخیره کنی
             print(f"Error sending reservation email: {e}")
-
-
-class AdminReservationListView(generics.ListAPIView):
-    queryset = Reservation.objects.select_related("user")
-    serializer_class = ReservationSerializer
-    permission_classes = [IsAdminUser]
-
-
-class DeleteResevationAPIView(generics.DestroyAPIView):
-    queryset = Reservation.objects.select_related("user")
-    permission_classes = [IsAdminUser]
-    serializer_class = ReservationSerializer
-
-
-class GetReservationCustomerListAPIView(generics.ListAPIView):
-    serializer_class = CustomerReservationSerializer
-    permission_classes = [IsAuthenticated]
-    
-
-    def get_queryset(self):
-        return Reservation.objects.filter(user=self.request.user)
